# refactor: optimize svg output