--- a/lib/calc-text-width.js
+++ b/lib/calc-text-width.js
@@ -1,10 +1,10 @@
+const round = require('./round')
 const widthsVerdana11 = require('./widths-verdana-11.json')
 const astralRegex = require('unicode-astral-regex')
 
 const calcWidth = (charWidthTable) => {
-  const SCALE = 10 // Prevent results like 60.599999999999994
-  const widthTable = charWidthTable.map(w => Math.round(w * SCALE))
-  widthTable[64] = widthTable[64] + 6 // Slightly increase width of "@" by 0.6px
+  const widthTable = charWidthTable.map(w => round(w))
+  widthTable[64] = widthTable[64] + 0.6 // Slightly increase width of "@" by 0.6px
 
   return (text, astral) => {
     if (astral) text = text.match(astralRegex) || []
@@ -16,7 +16,7 @@ const calcWidth = (charWidthTable) => {
       code = text[i].charCodeAt()
       total += widthTable[code < 127 ? code : 64] // Width as "@" for overflows
     }
-    return total / SCALE
+    return total
   }
 }