--- a/lib/index.js
+++ b/lib/index.js
@@ -1,5 +1,6 @@
-const calcWidth = require('./calc-text-width.js').Verdana11
-const colorPresets = require('./color-presets.js')
+const calcWidth = require('./calc-text-width').Verdana11
+const colorPresets = require('./color-presets')
+const round = require('./round')
 
 module.exports = ({ subject, status, color, style, emoji, icon, iconWidth = 13 }) => {
   typeAssert(typeof subject === 'string', '<subject> must be string')
@@ -9,8 +10,8 @@ module.exports = ({ subject, status, color, style, emoji, icon, iconWidth = 13 }
   const iconSpanWidth = icon ? (subject.length ? iconWidth + 4 : iconWidth - 2) : 0
   const sbTextWidth = calcWidth(subject, emoji)
   const stTextWidth = calcWidth(status, emoji)
-  const sbRectWidth = sbTextWidth + 10.2 + iconSpanWidth
-  const stRectWidth = stTextWidth + 10
+  const sbRectWidth = round(sbTextWidth + 10.2 + iconSpanWidth)
+  const stRectWidth = round(stTextWidth + 10)
   const width = sbRectWidth + stRectWidth
 
   if (style === 'flat') {