--- a/lib/index.js
+++ b/lib/index.js
@@ -17,18 +17,16 @@ module.exports = ({ subject, status, color, style, emoji, icon, iconWidth = 13 }
 
   if (style === 'flat') {
     return `
-        <g>
-          <rect width="${sbRectWidth}" height="20" fill="#555"/>
-          <rect x="${sbRectWidth}" width="${stRectWidth}" height="20" fill="#${color}"/>
-        </g>
-        <g fill="#fff" text-anchor="start" font-family="Verdana,DejaVu Sans,sans-serif" font-size="11">
-          <text x="${icon ? '22.5' : '6.3'}" y="14.8" textLength="${sbTextWidth}" fill="#000" opacity="0.1">${subject}</text>
       <svg width="${width}" height="20" xmlns="http://www.w3.org/2000/svg"${xlink}>
+        <path fill="#555" d="M0 0h${sbRectWidth}v20H0z"/>
+        <path fill="#${color}" d="M${sbRectWidth} 0h${stRectWidth}v20H${sbRectWidth}z"/>
+        <g fill="#fff" font-family="Verdana,DejaVu Sans,sans-serif" font-size="11">
+          <text x="${icon ? '22.5' : '6.3'}" y="14.8" textLength="${sbTextWidth}" fill="#000" opacity=".1">${subject}</text>
           <text x="${icon ? '21.5' : '5.3'}" y="13.8" textLength="${sbTextWidth}">${subject}</text>
-          <text x="${sbRectWidth + 5.5}" y="14.8" fill="#000" opacity="0.1" textLength="${stTextWidth}">${status}</text>
+          <text x="${sbRectWidth + 5.5}" y="14.8" textLength="${stTextWidth}" fill="#000" opacity=".1">${status}</text>
           <text x="${sbRectWidth + 4.5}" y="13.8" textLength="${stTextWidth}">${status}</text>
         </g>
-        ${icon ? `<image x="3.8" y="3.4" width="${iconWidth}" height="13.2" xlink:href="${icon}"/>` : ''}
+        ${icon ? `<image width="${iconWidth}" height="13.2" x="3.8" y="3.4" xlink:href="${icon}"/>` : ''}
       </svg>
     `
   }
@@ -39,19 +37,19 @@ module.exports = ({ subject, status, color, style, emoji, icon, iconWidth = 13 }
         <stop offset="0" stop-color="#EEE" stop-opacity=".1"/>
         <stop offset="1" stop-opacity=".1"/>
       </linearGradient>
-      <mask id="m"><rect width="${width}" height="20" rx="3" fill="#FFF"/></mask>
+      <mask id="m"><rect width="${width}" height="20" fill="#FFF" rx="3"/></mask>
       <g mask="url(#m)">
-        <rect width="${sbRectWidth}" height="20" fill="#555"/>
-        <rect x="${sbRectWidth}" width="${stRectWidth}" height="20" fill="#${color}"/>
-        <rect width="${width}" height="20" fill="url(#a)"/>
+        <path fill="#555" d="M0 0h${sbRectWidth}v20H0z"/>
+        <path fill="#${color}" d="M${sbRectWidth} 0h${stRectWidth}v20H${sbRectWidth}z"/>
+        <path fill="url(#a)" d="M0 0h${width}v20H0z"/>
       </g>
-      <g fill="#fff" text-anchor="start" font-family="Verdana,DejaVu Sans,sans-serif" font-size="11">
-        <text x="${icon ? '22.5' : '6.5'}" y="14.8" textLength="${sbTextWidth}" fill="#000" opacity="0.25">${subject}</text>
+      <g fill="#fff" font-family="Verdana,DejaVu Sans,sans-serif" font-size="11">
+        <text x="${icon ? '22.5' : '6.5'}" y="14.8" textLength="${sbTextWidth}" fill="#000" opacity=".25">${subject}</text>
         <text x="${icon ? '21.5' : '5.5'}" y="13.8" textLength="${sbTextWidth}">${subject}</text>
-        <text x="${sbRectWidth + 5.5}" y="14.8" fill="#000" opacity="0.25" textLength="${stTextWidth}">${status}</text>
+        <text x="${sbRectWidth + 5.5}" y="14.8" textLength="${stTextWidth}" fill="#000" opacity=".25">${status}</text>
         <text x="${sbRectWidth + 4.5}" y="13.8" textLength="${stTextWidth}">${status}</text>
       </g>
-      ${icon ? `<image x="3.9" y="3.5" width="${iconWidth}" height="13" xlink:href="${icon}"/>` : ''}
+      ${icon ? `<image width="${iconWidth}" height="13" x="3.9" y="3.5" xlink:href="${icon}"/>` : ''}
     </svg>
   `
 }