--- a/lib/round.js
+++ b/lib/round.js
@@ -1,0 +1,2 @@
+// Converts results like 60.599999999999994 to 60.6
+module.exports = value => Math.round(value * 10) / 10