--- a/tap-snapshots/test-round.spec.js-TAP.test.js
+++ b/tap-snapshots/test-round.spec.js-TAP.test.js
@@ -1,0 +1,14 @@
+/* IMPORTANT
+ * This snapshot file is auto-generated, but designed for humans.
+ * It should be checked into source control and tracked carefully.
+ * Re-generate by setting TAP_SNAPSHOT=1 and running tests.
+ * Make sure to inspect the output below.  Do not ignore changes!
+ */
+'use strict'
+exports[`test/round.spec.js TAP rounds integer to zero decimal places > result is correct 1`] = `
+1
+`
+
+exports[`test/round.spec.js TAP rounds decimal to one decimal place > result is correct 1`] = `
+1.6
+`