… tap-snapshots/test-index.spec.js-TAP.test.js