#### Changes

- Extends rounding to the rectangle widths instead of just the text width.
- Removes `xmlns:xlink` from the SVG when no icon is present
- Reordered some properties to be consistent
- Converted `rect` to `path`, removing the need for 2 groups in flat badges
   ```diff
   -<g><rect width="35.1" height="20" fill="#555"/><rect x="35.1" width="45.5" height="20" fill="#0BC"/></g>
   +<path fill="#555" d="M0 0h35.1v20H0z"/><path fill="#08C" d="M35.1 0h45.5v20H35.1z"/>
   ```

Based on the first snapshot, with this change unnecessary whitespace has gone from 14% to 20% of the badge size (~150 bytes) and the badge size reduced by 140 bytes.

Further reductions can be made in [badgen-service](https://github.com/amio/badgen-service) with gzip/brotli support:
- gzip reduces the size by 54% (~500 bytes). #3 (but at prepublish, not runtime) would only save an extra 20 bytes on the output.
- brotli reduces the badge size by 61% (~570 bytes) with #3 saving the same amount.

#### Benchmarks

##### PR

```
[classic] style, long params x 797,333 ops/sec ±1.87% (89 runs sampled)
[classic] style, full params x 885,457 ops/sec ±1.48% (88 runs sampled)
   [flat] style, long params x 528,192 ops/sec ±1.13% (91 runs sampled)
   [flat] style, full params x 464,142 ops/sec ±4.81% (74 runs sampled)
[classic] style, with emoji  x 491,251 ops/sec ±1.43% (88 runs sampled)
[classic] style, with icon   x 718,199 ops/sec ±1.20% (89 runs sampled)
   [flat] style, with emoji  x 330,065 ops/sec ±1.11% (91 runs sampled)
   [flat] style, with icon   x 431,363 ops/sec ±0.85% (90 runs sampled)
```

##### Master

```
[classic] style, long params x 654,538 ops/sec ±2.61% (78 runs sampled)
[classic] style, full params x 766,651 ops/sec ±1.55% (85 runs sampled)
   [flat] style, long params x 449,018 ops/sec ±2.14% (82 runs sampled)
   [flat] style, full params x 470,980 ops/sec ±1.78% (83 runs sampled)
[classic] style, with emoji  x 466,423 ops/sec ±2.08% (83 runs sampled)
[classic] style, with icon   x 647,043 ops/sec ±1.63% (82 runs sampled)
```