@amio 2018-08-17
Thanks! @wopian 

+1 on removeing `xmlns:xlink` & reordering properties 🎉

And some thoughts on rest changes:

- `rect` vs `path`. I use `path` at frist, but later change to `rect` for better readability, and 1 less string concat:
```diff
- <path fill="#${color}" d="M${sbRectWidth} 0h${stRectWidth}v20H${sbRectWidth}z"/>
+ <rect fill="#${color}" x="${sbRectWidth}" width="${stRectWidth}" height="20"/>
```

- Rounding number issue. +1 on this, but it seems still [not fully covered](https://github.com/amio/badgen/pull/10/files#diff-96d822e7753e8c518bd1c218082688b0R99). I have another idea for this, generate all units in 10x, and use `viewBox` to do the scale. Get rid of all the rounding maths. 💪