@amio 2018-08-22
@wopian I've done rest changes in [v2.3.0](https://github.com/amio/badgen/releases/tag/v2.3.0), this PR can be closed. Thanks!