@wopian 2018-08-22
Sorry for the lack of response, did this PR just before leaving for a week holiday 😞

I still much prefer the path syntax over height/width though.