@amio 2018-08-23
Not at all @wopian , happy holiday!

I found there's one more benefit of using path, we can use ['arc path'](https://stackoverflow.com/a/38118843/2252410) to get rid of that mask for rounded corner:
```svg
<mask id="m"><rect width="${width}" height="20" fill="#FFF" rx="3"/></mask>
```